BE:refactor: create enum class for cookie & token
