+++ backend/member/src/main/java/com/nekarak8s/member/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.nekarak8s.member.common.config;
+package com.nekarak8s.member.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
