+++ backend/member/src/main/java/com/nekarak8s/member/util/cookie/CookieUtils.java
@@ -1,5 +1,6 @@
 package com.nekarak8s.member.util.cookie;
 
+import com.nekarak8s.member.common.ExpTime;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
@@ -9,14 +10,12 @@ import javax.servlet.http.HttpServletResponse;
 @Component
 @Slf4j
 public class CookieUtils {
-
-    private final String COOKIE_NAME = "gallery_cookie";
+    private static final String COOKIE_NAME = "gallery_cookie";
 
     public void addCookie(HttpServletResponse response, String token) {
-
         Cookie cookie = new Cookie(COOKIE_NAME, token);
-        cookie.setSecure(false); // https 사용시 true로 변경
-        cookie.setMaxAge(600 * 4); // 40분 = 2400초  (초 단위)
+        cookie.setSecure(false); // https 사용시 true 로 변경
+        cookie.setMaxAge(8 * ExpTime.SECONDS_PER_HOUR.getValue());
         cookie.setPath("/");
         cookie.setHttpOnly(false); // dev
         //cookie.setHttpOnly(true); // prod
