+++ backend/member/src/main/java/com/nekarak8s/member/util/jwt/JwtUtils.java
@@ -3,6 +3,7 @@ package com.nekarak8s.member.util.jwt;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
+import com.nekarak8s.member.common.ExpTime;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.InitializingBean;
@@ -14,26 +15,21 @@ import java.util.Date;
 @Slf4j
 @RequiredArgsConstructor
 public class JwtUtils implements InitializingBean {
-
     private static final String ROLE_CLAIM_KEY = "role";
-
     private final JwtProperties jwtProperties;
 
     private Algorithm algorithm;
     private JWTVerifier jwtVerifier;
 
     @Override
-    public void afterPropertiesSet() throws Exception {
+    public void afterPropertiesSet() {
         this.algorithm = Algorithm.HMAC512(jwtProperties.getSecret());
         this.jwtVerifier = JWT.require(algorithm).acceptLeeway(5).build(); // 여유 시간 5분 설정
     }
 
     public String generate(TokenMember member) {
         Date now = new Date();
-        log.debug("갤러리 accessToken 생성 시작");
-        log.debug("[<-->]토큰 만료 시간(ms) : {}ms", jwtProperties.getExpiration() * 60 * 1000);
-        log.debug("[<-->]토큰 만료 시간(분) : {}분", jwtProperties.getExpiration());
-        Date expiresAt = new Date(now.getTime() + jwtProperties.getExpiration() * 60 * 1000);
+        Date expiresAt = new Date(now.getTime() + jwtProperties.getExpiration() * ExpTime.MILLISECONDS_PER_HOUR.getValue());
 
         return JWT.create()
                 .withSubject(member.getId())
